Vector: return Matrices for comb* operations.
Issue #4344.

	* include/urbi/object/vector.hh: Use size_t instead of unsigned int.
	(vector_type, matrix_type): New.
	(combAdd, combSub, combDiv, combMul): Take arguments by const ref.
	Return a matrix instead of a vector.
	(selfCombAdd, selfCombSub, selfCombDiv, selfCombMul): Remove.
	* include/urbi/object/vector.hxx, src/object/vector.cc: Adjust.

+++ include/urbi/object/vector.hxx
@@ -93,7 +93,7 @@ namespace urbi
         return this->Object::operator<(bb);
       if (value_.size() != b->value_.size())
         return value_.size() < b->value_.size();
-      for (unsigned int i=0; i<value_.size(); ++i)
+      for (size_t i = 0; i < value_.size(); ++i)
         if (value_(i) != b->value_(i))
           return value_(i) < b->value_(i);
       return false;
